refactor: implement gradle
