+++ app/src/main/java/sap/ass01/bbom/AddEBikeDialog.java
@@ -1,5 +1,4 @@
 package sap.ass01.bbom;
-
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
