+++ app/src/main/java/sap/ass01/layered/domain/EBike.java
